+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `erlang:spawn_monitor/1`, `erlang:spawn_monitor/3`
 - Added `lists:dropwhile/2`.
 - Support for `float/1` BIF.
+- Added `erlang:get/0` and `erlang:erase/0`.
 
 ### Fixed
 - ESP32: improved sntp sync speed from a cold boot.